zzogl-pg: A few Windows changes, and fix an incorrect variable type.
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@3450 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/zzogl-pg/opengl/GS.h
@@ -645,7 +645,7 @@ typedef struct
 			  ((a & 0xff) << 24);
 	}
 	
-	void add_vertex(u16 x, u16 y, u32 z, float f)
+	void add_vertex(u16 x, u16 y, u32 z, u16 f)
 	{
 		vertexregs.x = x;
 		vertexregs.y = y;
